Issue #66 - Getting a Delegate to callback from C++ into Swift
So, it turns out that there is a way to callback from C++ into
Swift. The key is that Obj-C++ can silently wrap a Swift object in a
C++ class. This patch introduces a few delegates for a few tasks. Once
this patch has passed review, I'm planning to continue work with other
delegates.

To make this work, I had to tweak message-passing to the event
loop. The event loop now supports enqueing of messages. We might find
out at a later stage that the semantics of this enqueuing need some
clarification.

Also, we have our first unit test \o/. And it's async.